refactor: utils and add function for getting date range
