+++ hrms/tests/test_utils.py
@@ -1,5 +1,37 @@
 import frappe
-from frappe.utils import add_months, get_first_day, get_last_day, getdate
+from frappe.utils import add_months, get_first_day, get_last_day, getdate, now_datetime
+
+from erpnext.setup.utils import enable_all_roles_and_domains
+
+
+def before_tests():
+	frappe.clear_cache()
+	# complete setup if missing
+	from frappe.desk.page.setup_wizard.setup_wizard import setup_complete
+
+	year = now_datetime().year
+	if not frappe.get_list("Company"):
+		setup_complete(
+			{
+				"currency": "INR",
+				"full_name": "Test User",
+				"company_name": "Wind Power LLC",
+				"timezone": "Asia/Kolkata",
+				"company_abbr": "WP",
+				"industry": "Manufacturing",
+				"country": "India",
+				"fy_start_date": f"{year}-01-01",
+				"fy_end_date": f"{year}-12-31",
+				"language": "english",
+				"company_tagline": "Testing",
+				"email": "test@erpnext.com",
+				"password": "test",
+				"chart_of_accounts": "Standard",
+			}
+		)
+
+	enable_all_roles_and_domains()
+	frappe.db.commit()  # nosemgrep
 
 
 def get_first_sunday(holiday_list="Salary Slip Test Holiday List", for_date=None):
