--- hrms/utils.py
@@ -1,60 +1,0 @@
-import requests
-
-import frappe
-from frappe.utils import now_datetime
-
-from erpnext.setup.utils import enable_all_roles_and_domains
-
-country_info = {}
-
-
-@frappe.whitelist(allow_guest=True)
-def get_country(fields=None):
-	global country_info
-	ip = frappe.local.request_ip
-
-	if ip not in country_info:
-		fields = ["countryCode", "country", "regionName", "city"]
-		res = requests.get(
-			"https://pro.ip-api.com/json/{ip}?key={key}&fields={fields}".format(
-				ip=ip, key=frappe.conf.get("ip-api-key"), fields=",".join(fields)
-			)
-		)
-
-		try:
-			country_info[ip] = res.json()
-
-		except Exception:
-			country_info[ip] = {}
-
-	return country_info[ip]
-
-
-def before_tests():
-	frappe.clear_cache()
-	# complete setup if missing
-	from frappe.desk.page.setup_wizard.setup_wizard import setup_complete
-
-	year = now_datetime().year
-	if not frappe.get_list("Company"):
-		setup_complete(
-			{
-				"currency": "INR",
-				"full_name": "Test User",
-				"company_name": "Wind Power LLC",
-				"timezone": "Asia/Kolkata",
-				"company_abbr": "WP",
-				"industry": "Manufacturing",
-				"country": "India",
-				"fy_start_date": f"{year}-01-01",
-				"fy_end_date": f"{year}-12-31",
-				"language": "english",
-				"company_tagline": "Testing",
-				"email": "test@erpnext.com",
-				"password": "test",
-				"chart_of_accounts": "Standard",
-			}
-		)
-
-	enable_all_roles_and_domains()
-	frappe.db.commit()  # nosemgrep
