+++ hrms/utils/__init__.py
@@ -1,0 +1,34 @@
+import requests
+
+import frappe
+from frappe.utils import add_days, date_diff
+
+country_info = {}
+
+
+@frappe.whitelist(allow_guest=True)
+def get_country(fields=None):
+	global country_info
+	ip = frappe.local.request_ip
+
+	if ip not in country_info:
+		fields = ["countryCode", "country", "regionName", "city"]
+		res = requests.get(
+			"https://pro.ip-api.com/json/{ip}?key={key}&fields={fields}".format(
+				ip=ip, key=frappe.conf.get("ip-api-key"), fields=",".join(fields)
+			)
+		)
+
+		try:
+			country_info[ip] = res.json()
+
+		except Exception:
+			country_info[ip] = {}
+
+	return country_info[ip]
+
+
+def get_date_range(start_date: str, end_date: str) -> list[str]:
+	"""returns list of dates between start and end dates"""
+	no_of_days = date_diff(end_date, start_date) + 1
+	return [add_days(start_date, i) for i in range(no_of_days)]
